feat: 없는 Article을 삭제하려고 할 때 EntityNotFoundException와 예외 메시지를 반환하도록 하는 기능 구현

+++ src/main/java/toyboard/yoon/service/ArticleService.java
@@ -72,6 +72,12 @@ public class ArticleService {
 
     @Transactional
     public void deleteArticle(Long articleId) {
+        Optional<Article> article = articleRepository.findById(articleId);
+
+        if(article.isEmpty()) {
+            throw new EntityNotFoundException(String.format("Article Id '%d'가 존재하지 않습니다.", articleId));
+        }
+
         articleRepository.deleteById(articleId);
     }
 }
